chore: add tests for RedisStorage2
